pyproject.toml
[project]
name = "icms-antecipado-pi"
version = "0.1.0"
description = "ICMS Antecipado — automação SEFAZ-PI (DAR Web)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
icms_pi = "icms_pi.gui_app:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["atc*", "icms_pi*"]

[tool.uv]
dev-dependencies = []
